server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Thierry <thierrytajesdossantos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
python-dotenv = "^1.0.0"
python-jose = "^3.3.0"
sqlmodel = "^0.0.8"
dynaconf = "^3.1.12"
bcrypt = "^4.0.1"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.22.0"
black = "^23.3.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
